requirements.txt
Flask==0.10.1
flask-socketio==0.4.0
Flask-Cors==1.9.0
Jinja2==2.7.2
Markdown==2.4.1
numpy==1.9.0
Twisted==13.2.0
Unidecode==0.04.16
Werkzeug==0.9.4
pyserial==2.6
